requirements.txt
absl-py==0.10.0
astunparse==1.6.3
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2021.5.30
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
cytoolz==0.10.1
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.7
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.11.2
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
protobuf==3.13.0
ptyprocess==0.6.0
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacred==0.8.1
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.6.0
torchfile==0.1.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
urllib3==1.25.10
visdom==0.1.8.9
visdom-logger @ git+git://github.com/luizgh/visdom_logger@3c0afb18321ef65382d0baf8866772c6a62f15a2
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
